Upgrade test runner to derive from DiscoveryRunner.
