+++ evennia/settings_default.py
@@ -610,7 +610,7 @@ INSTALLED_APPS = (
 # The user profile extends the User object with more functionality;
 # This should usually not be changed.
 AUTH_USER_MODEL = "players.PlayerDB"
-#AUTH_PROFILE_MODULE = "players.PlayerDB"
+
 # Use a custom test runner that just tests Evennia-specific apps.
 TEST_RUNNER = 'evennia.server.tests.EvenniaTestSuiteRunner'
