+++ evennia/utils/test_resources.py
